update CMakeFile.txt patch for r2382
Without this change, compiling still works, but warns about  HUNK#2..4 being applied at an offset. 

I'm not sure what happens to this file as I update through the GUI and it seems to be sensitive to having dos newlines, which show up as ^M in linux. as I cannot see these in the web gui of GitHub, I'm not sure if they will be included, so this might break compilation if not included. On my PC I do a unix2dos to fix the line endings.